requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-groq
langchain-community
pydantic
transformers
requests
gradio
